Inflated palette in mode 15

+++ Pokitto/POKITTO_HW/HWLCD.h
@@ -80,7 +80,7 @@ extern void lcdRefreshMode3(uint8_t *, uint16_t*);
 extern void lcdRefreshModeGBC(uint8_t *, uint16_t*);
 extern void lcdRefreshMode13(uint8_t *, uint16_t*, uint8_t);
  
-extern void lcdRefreshMode15(uint16_t*, uint8_t*);
+extern void lcdRefreshMode15(uint32_t*, uint8_t*);
  
 
 /** Update LCD from 1-bit tile mode */
